# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API completa para gestão de equinos com Tokenização RWA (Real World Assets). Inclui: autenticação, gestão de equinos, tokenização, leilões, exames laboratoriais, nutrição com IA, treinamento, rankings, financeiro e muito mais.",
        "title": "Equinoid API - Sistema Perfeito",
        "termsOfService": "https://equinoid.com/terms",
        "contact": {
            "name": "Equinoid Support",
            "url": "https://equinoid.com/support",
            "email": "suporte@equinoid.com"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://equinoid.com/license"
        },
        "version": "2.0"
    },
    "host": "seahorse-app-28du8.ondigitalocean.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/equinos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os equinos com filtros e paginação",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Listar equinos",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Itens por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar por nome ou Equinoid",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por raça",
                        "name": "raca",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por proprietário",
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por veterinário",
                        "name": "veterinario_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo equino no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Criar novo equino",
                "parameters": [
                    {
                        "description": "Dados do equino",
                        "name": "equino",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEquinoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/equinos/{equinoid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados completos de um equino",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Buscar equino por Equinoid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um equino existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Atualizar equino",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "equino",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEquinoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um equino do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Deletar equino",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/equinos/{equinoid}/transferir": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfere a propriedade de um equino para outro usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equinos"
                ],
                "summary": "Transferir propriedade de equino",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID do novo proprietário",
                        "name": "transfer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "novo_proprietario_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateEquinoRequest": {
            "type": "object",
            "required": [
                "data_nascimento",
                "microchip_id",
                "nome",
                "pais_origem",
                "pelagem",
                "proprietario_id",
                "raca",
                "sexo"
            ],
            "properties": {
                "data_nascimento": {
                    "type": "string"
                },
                "external_ids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExternalID"
                    }
                },
                "foto_perfil": {
                    "type": "string"
                },
                "genitor_equinoid": {
                    "type": "string"
                },
                "genitora_equinoid": {
                    "type": "string"
                },
                "microchip_id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "pais_origem": {
                    "type": "string"
                },
                "pelagem": {
                    "type": "string"
                },
                "propriedade_id": {
                    "type": "integer"
                },
                "proprietario_id": {
                    "type": "integer"
                },
                "raca": {
                    "type": "string"
                },
                "sexo": {
                    "$ref": "#/definitions/models.SexoEquino"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.ExternalID": {
            "type": "object",
            "required": [
                "id",
                "sistema"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "sistema": {
                    "type": "string"
                }
            }
        },
...
        "models.SexoEquino": {
            "type": "string",
            "enum": [
                "macho",
                "femea"
            ],
            "x-enum-varnames": [
                "SexoMacho",
                "SexoFemea"
            ]
        },
...
        "models.StatusEquino": {
            "type": "string",
            "enum": [
                "ativo",
                "inativo",
                "vendido",
                "falecido",
                "transferir"
            ],
            "x-enum-varnames": [
                "StatusAtivo",
                "StatusInativo",
                "StatusVendido",
                "StatusFalecido",
                "StatusTransferir"
            ]
        },
...
        "models.UpdateEquinoRequest": {
            "type": "object",
            "properties": {
                "external_ids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExternalID"
                    }
                },
                "foto_perfil": {
                    "type": "string"
                },
                "fotos_galeria": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "pais_origem": {
                    "type": "string"
                },
                "pelagem": {
                    "type": "string"
                },
                "propriedade_id": {
                    "type": "integer"
                },
                "raca": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StatusEquino"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}